Add a starting_position attribute to ChessPiece(s)
In order to avoid issues and coupling, each piece will get a starting
position attribute, simplifying the algorithm to determine if it has
moved or not.

It becomes the domain of the piece to know if it has moved or not,
reducing coupling and issues throughout the rest of the program.